requirements.txt
bs4
Pillow==4.0.0
requests==2.13.0
